requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
dj-static==0.0.6
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
mongoengine==0.24.2
numpy==1.23.2
pandas==1.4.3
PyJWT==2.4.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
static3==0.7.0
threadpoolctl==3.1.0
tzdata==2022.2
urllib3==1.26.12
waitress==2.1.2
xgboost==1.6.2
